On partial read/write, keep reading.
This avoids rsync errors and avoids data loss with cp, due to
failing to copy the last byte of 0.1953125% of files.

Unless direct_io is set, fuse expects a full read/write.

As a workaround for a device hanging bug, if the length of a file
on a Samsung device is 500 bytes modulo 512 bytes, libmtp returns a
partial read omitting the last byte which must be requested
separately.

Fixes #85.